backend/pyproject.toml
[project]
name = "bookfield"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
